+++ docs/CHANGELOG.md
@@ -15,6 +15,14 @@ Changes are grouped as follows:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [0.26.0] - 29-10-23
+
+### Added
+- Added rules2excel rules exporter. Now users can export rules from TransformationRules object to excel file.
+- Added rules generator step from arbitrary object (must be in json or yaml formats)
+- Added eperimental rules parser from OpenApi/Swagger specification. Rules generates based on schema part of OpenApi specification.
+- Added version , source and docs_urs metadata to Steps class.
+
 ## [0.25.9] - 30-09-23
 ### Fixed
 * Loading `neat` from environment variables, the variable `NEAT_LOAD_EXAMPLES` would always return `true`